Minor bug fixes

+++ src/Vehicles/Car.java
@@ -7,7 +7,7 @@ public class Car extends Vehicle
 {
     int numberOfDoors;
 
-    public Car(String type, String model, int buildYear, int numberOfDoors, Tile[][] map, Tile entryPoint, Tile exitPoint, int allowedSpeed, RailwayCrossing railwayCrossing)
+    public Car(String type, String model, int buildYear, int numberOfDoors, Tile[][] map, Tile entryPoint, Tile exitPoint, double allowedSpeed, RailwayCrossing railwayCrossing)
     {
         super(type, model, buildYear, map, entryPoint, exitPoint, allowedSpeed, railwayCrossing);
         this.numberOfDoors = numberOfDoors;
