+++ src/Vehicles/Truck.java
@@ -7,7 +7,7 @@ public class Truck extends Vehicle
 {
     double weightAllowed;
 
-    public Truck(String type, String model, int buildYear, double weightAllowed, Tile[][] map, Tile entryPoint, Tile exitPoint, int allowedSpeed, RailwayCrossing railwayCrossing)
+    public Truck(String type, String model, int buildYear, double weightAllowed, Tile[][] map, Tile entryPoint, Tile exitPoint, double allowedSpeed, RailwayCrossing railwayCrossing)
     {
         super(type, model, buildYear, map,entryPoint,exitPoint, allowedSpeed, railwayCrossing);
         this.weightAllowed = weightAllowed;
